Refactor : GeometryUtil 클래스 하나로 통일
