--- parker/src/main/java/com/si9nal/parker/map/dto/response/Location.java
@@ -1,15 +1,0 @@
-package com.si9nal.parker.map.dto.response;
-
-import lombok.Getter;
-
-@Getter
-public class Location {
-
-    private Double latitude;
-    private Double longitude;
-
-    public Location(Double latitude, Double longitude) {
-        this.latitude = latitude;
-        this.longitude = longitude;
-    }
-}
